requirements.txt
# Core Dependencies
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-neo4j>=0.1.0
graphrag>=0.3.0

# Azure OpenAI
openai>=1.0.0

# Data Processing
pandas>=2.0.0
openpyxl>=3.1.0

# Neo4j
neo4j>=5.0.0

# Environment
python-dotenv>=1.0.0

# Report Export
python-docx>=1.1.0
reportlab>=4.0.0

# Utilities
pydantic>=2.0.0
typing-extensions>=4.5.0
